refactor redux login stuff into header component, moving state upwards
