+++ src/components/header.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { Dispatch } from 'react';
+import { connect } from 'react-redux';
 import { Link } from 'gatsby';
 
 // Material UI imports
@@ -14,6 +15,8 @@ import MenuIcon from '@material-ui/icons/Menu';
 
 import SigninDialog from './signinDialog';
 import Navbar from './navbar';
+
+import { loginAction, logoutAction } from '../state/actions/auth';
 
 import {
   AppToolBarBackgroundColor,
@@ -39,7 +42,11 @@ const useStyles = makeStyles(() =>
   })
 );
 
-const Header: React.FC<HeaderProps> = ({ siteTitle = '' }: HeaderProps) => {
+const Header: React.FC<HeaderProps> = ({
+  siteTitle = '',
+  signinAction,
+  signoutAction,
+}: HeaderProps) => {
   const classes = useStyles();
 
   const [signinDialogOpen, setSigninDialogOpen] = React.useState(false);
@@ -85,14 +92,42 @@ const Header: React.FC<HeaderProps> = ({ siteTitle = '' }: HeaderProps) => {
           </Button>
         </Toolbar>
       </AppBar>
-      <SigninDialog open={signinDialogOpen} onClose={handleCloseSignDialog} />
+      <SigninDialog
+        open={signinDialogOpen}
+        onClose={handleCloseSignDialog}
+        onGoogleLoginButtonClick={() => signinAction(true)}
+      />
       <Navbar open={navbarOpen} onClose={handleCloseNavbar} />
     </header>
   );
 };
 
-interface HeaderProps {
+export interface HeaderProps {
   siteTitle: string;
+  isLoggedIn: boolean;
+  signinAction: (withGoogle: boolean) => void;
+  signoutAction: () => void;
 }
 
-export default Header;
+export interface HeaderStateMapping {
+  isLoggedIn: boolean;
+}
+
+export interface HeaderDispatchMapping {
+  signinAction: (withGoogle: boolean) => void;
+  signoutAction: () => void;
+}
+
+const mapStateToProps = (state: HeaderStateMapping): HeaderStateMapping => ({
+  isLoggedIn: state.isLoggedIn,
+});
+
+const mapDispatchToProps = (
+  // TODO: Fix this any ts type
+  dispatch: Dispatch<any>
+): HeaderDispatchMapping => ({
+  signinAction: (withGoogle: boolean) => dispatch(loginAction(withGoogle)),
+  signoutAction: () => dispatch(logoutAction()),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(Header);
