+++ src/components/signinDialog.tsx
@@ -1,5 +1,4 @@
 import React, { Dispatch } from 'react';
-import { connect } from 'react-redux';
 
 // Material UI imports
 import {
@@ -13,14 +12,10 @@ import {
 
 import GoogleButton from 'react-google-button/dist/react-google-button';
 
-import { AuthAction, loginAction, logoutAction } from '../state/actions/auth';
-
 const SigninDialog: React.FC<SigninDialogProps> = ({
   open,
   onClose,
-  isLoggedIn,
-  signinAction,
-  signoutAction,
+  onGoogleLoginButtonClick,
 }) => {
   return (
     <Dialog onClose={onClose} aria-labelledby="simple-dialog-title" open={open}>
@@ -40,7 +35,7 @@ const SigninDialog: React.FC<SigninDialogProps> = ({
           </Button>
         </ListItem>
         <ListItem>
-          <GoogleButton onClick={() => signinAction(true)} />
+          <GoogleButton onClick={onGoogleLoginButtonClick} />
         </ListItem>
       </List>
     </Dialog>
@@ -50,32 +45,7 @@ const SigninDialog: React.FC<SigninDialogProps> = ({
 export interface SigninDialogProps {
   open: boolean;
   onClose: () => void;
-  isLoggedIn: boolean;
-  signinAction: (withGoogle: boolean) => void;
-  signoutAction: () => void;
+  onGoogleLoginButtonClick: () => void;
 }
 
-export interface SigninDialogStateMapping {
-  isLoggedIn: boolean;
-}
-
-export interface SigninDialogDispatchMapping {
-  signinAction: (withGoogle: boolean) => void;
-  signoutAction: () => void;
-}
-
-const mapStateToProps = (
-  state: SigninDialogStateMapping
-): SigninDialogStateMapping => ({
-  isLoggedIn: state.isLoggedIn,
-});
-
-const mapDispatchToProps = (
-  // TODO: Fix this any ts type
-  dispatch: Dispatch<any>
-): SigninDialogDispatchMapping => ({
-  signinAction: (withGoogle: boolean) => dispatch(loginAction(withGoogle)),
-  signoutAction: () => dispatch(logoutAction()),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(SigninDialog);
+export default SigninDialog;
